requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
pip~=23.2.1
pillow~=10.2.0
docutils~=0.20.1
setuptools~=69.0.3
cffi~=1.16.0
pyinstaller~=6.3.0
future~=0.18.3